requirements.txt
django-heroku
Django==2.1.11
gunicorn 
psycopg2
django-mathfilters
requests
django chatterbot
chatterbot-corpus